+++ test/TestFighter.cpp
@@ -1,4 +1,4 @@
-// Copyright 2018 Florian Muecke. All rights reserved.
+// Copyright 2021 Florian Muecke. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE.txt file.
 
@@ -67,4 +67,22 @@ TEST_CASE("operator == does not match empty category")
 	REQUIRE_FALSE(GetDummy() == f1);
 	REQUIRE(f1 != GetDummy());
 	REQUIRE(GetDummy() != f1);
+}
+
+TEST_CASE("Matches matches empty club")
+{
+    auto f1 = GetDummy();
+    f1.club = "";
+    REQUIRE(f1.Matches(GetDummy()));
+    REQUIRE(GetDummy().Matches(f1));
+}
+
+TEST_CASE("operator == does not match empty club")
+{
+    auto f1 = GetDummy();
+    f1.club = "";
+    REQUIRE_FALSE(f1 == GetDummy());
+    REQUIRE_FALSE(GetDummy() == f1);
+    REQUIRE(f1 != GetDummy());
+    REQUIRE(GetDummy() != f1);
 }